tests/submission/Infraware/Forms/contents/Forms/Not Use/select_required.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>select_required</h3>
	</p>
	
	<li>When the select control is not selected, you cannot submit this form.</li>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action="?x=1"
		  name="input_form">
		<p>
		 <label for="unittype">Select unit type:</label>
		 <select id="unittype" name="unittype" multiple>
		  <option value="1"> Miner </option>
		  <option value="2"> Puffer </option>
		  <option value="3"> Snipey </option>
		  <option value="4"> Max </option>
		  <option value="5"> Firebot </option>
		 </select>
		</p>	
		<p><input type='submit' id='input_submit'></p>
	</form>

	<script>
		
		var select = document.getElementById("unittype");
		select.required = true;

		if (typeof(select.required) == "boolean") {
			if (select.required != true) {
				test(function() {
					assert_true(false, "required attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("required attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
